+++ src/decl/assert-decl.h
@@ -1,0 +1,4 @@
+static r_no_return
+void stop_non_list_type(r_obj* x,
+                        struct vctrs_arg* arg,
+                        struct r_lazy call);